move ownedbytes to own crate
fixes #1106
